refactor(quality): establish domain and repository boundaries
