--- yak-ops-business/yak-ops-business-quality/src/main/java/io/yak/ops/business/quality/execution/QualityRuntime.java
@@ -1,62 +1,0 @@
-package io.yak.ops.business.quality.execution;
-
-import io.yak.ops.common.enums.quality.QualityEnums.AlertLevel;
-import io.yak.ops.common.enums.quality.QualityEnums.ComparisonOperator;
-import io.yak.ops.common.enums.quality.QualityEnums.NotifyChannel;
-import io.yak.ops.common.enums.quality.QualityEnums.RuleFailureAction;
-import io.yak.ops.common.enums.quality.QualityEnums.RuleScope;
-import io.yak.ops.common.enums.quality.QualityEnums.RuleType;
-import java.math.BigDecimal;
-import java.util.List;
-
-public final class QualityRuntime {
-
-  private QualityRuntime() {}
-
-  public record MonitorSnapshot(
-      long id,
-      String name,
-      long dataSourceId,
-      String dataSourceName,
-      String databaseName,
-      String schemaName,
-      String tableName,
-      String whereClause,
-      String owner) {}
-
-  public record RuleSnapshot(
-      long id,
-      long templateId,
-      String templateCode,
-      String name,
-      RuleType ruleType,
-      RuleScope scope,
-      String dimension,
-      String columnName,
-      ComparisonOperator operator,
-      BigDecimal threshold,
-      BigDecimal thresholdEnd,
-      List<String> enumValues,
-      String customSql) {}
-
-  public record ExecutionJob(
-      long executionId,
-      String executionNo,
-      MonitorSnapshot monitor,
-      List<RuleSnapshot> rules,
-      RuleFailureAction ruleFailureAction,
-      boolean notifyEnabled,
-      NotifyChannel notifyChannel,
-      String notifyTarget,
-      AlertLevel alertLevel) {
-
-    public ExecutionJob(
-        long executionId,
-        String executionNo,
-        MonitorSnapshot monitor,
-        List<RuleSnapshot> rules) {
-      this(executionId, executionNo, monitor, rules, RuleFailureAction.CONTINUE,
-          false, NotifyChannel.MESSAGE, null, AlertLevel.WARNING);
-    }
-  }
-}
