+++ yak-ops-business/yak-ops-business-quality/src/main/java/io/yak/ops/business/quality/repository/QualityAlertRepository.java
@@ -1,0 +1,8 @@
+package io.yak.ops.business.quality.repository;
+
+import io.yak.ops.business.quality.domain.QualityDomain.AlertEventSpec;
+
+/** Persistence port for quality alert events. */
+public interface QualityAlertRepository {
+  void insertAlertEvent(AlertEventSpec alert);
+}
